moved single components into a new home folder
