+++ skizzy-fast-food/src/App.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 import '../src/scss/styles.scss'
-import Nav from './components/Nav'
-import Header from './components/Header'
-import Menu from './components/menu/Menu'
+import Nav from './components/home/Nav'
+import Header from './components/home/Header'
+import Menu from './components/home/Menu'
 import './App.css'
 
 function App() {
